fix: add repo route in another file and change pyproject.toml for the render
