+++ frontend/src/services/api.ts
@@ -1,14 +1,15 @@
 import axios from "axios"
 
-const baseURL = process.env.URL
+const baseURL = "https://amazed-infinitely-marten.ngrok-free.app"
 
-console.log("baseURL:",{baseURL})
+console.log("baseURL:",baseURL)
 
 const apiClient  = axios.create(
     {
         baseURL,
         headers:{
             'Content-Type': 'application/json',
+            'ngrok-skip-browser-warning': 'true'
         },
         withCredentials: true
     }
